Transferred from the Adobe Open Source Subversion Repository
Revision: 7288
Author:   chaase@adobe.com
Date:     2009-05-26T22:42:32.255273Z

Log Message: 
-----------
Fixed bug in Rotate3D introduced when it moved out from under Rotate, which used to be its superclass. The problem was that we were not re-initializing the motionPaths array of AnimateTransform every time the Rotate3D effect was played, so we would just end up adding more and more keyframes, which ended up giving conflicting animation information to the effect.
The fix was to pull in the same createInstance() override which is already in Move, Rotate, and Scale, and do the same initialization of motionPaths[] as we do in those other subclasses of AnimateTransform.

QE Notes: None
Doc Notes: None
Bugs: sdk-21170
Reviewer: Evtim
tests: checkintests mustella spark/effects/Rotate3D

Modified Paths:
--------------
/flex/sdk/trunk/frameworks/projects/flex4/src/spark/effects/Rotate3D.as

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1283733 13f79535-47bb-0310-9956-ffa450edef68