Transferred from the Adobe Open Source Subversion Repository
Revision: 3280
Author:   chaase@adobe.com
Date:     2008-09-20T00:27:13.071102Z

Log Message: 
-----------
Fix for SDK-16951. Resize was sometimes incorrectly attempting to animate properties that were set to NaN (default value) in the target. Fix is to check for NaN's for these from/to values before bothering to push an additional propertyValuesHolder item onto the list of properties to Animate.as

Reviewed by Glenn
tests: checkintests

Modified Paths:
--------------
/flex/sdk/trunk/frameworks/projects/flex4/src/flex/effects/effectClasses/ResizeInstance.as

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1284432 13f79535-47bb-0310-9956-ffa450edef68