Transferred from the Adobe Open Source Subversion Repository
Revision: 11798
Author:   jszeto@adobe.com
Date:     2009-11-13T22:01:46.197040Z

Log Message: 
-----------
Border bug fix

http://bugs.adobe.com/jira/browse/SDK-23824 - Border size does not update when a child element is included after a state change

The bug was that we were explicitly setting the width and height of the contentGroup in updateDisplayList. This meant that if content was added later, it was not forcing a resize of the contentGroup and thus the skin did not update. Using setActualSize does not work because it breaks constraints set on children. The solution is to use constraints on the contentGroup. 

QE notes: none
Doc notes: none
Bugs: SDK-23824
Reviewer: Kevin
Tests run: Border
Is noteworthy for integration: No

Modified Paths:
--------------
/flex/sdk/trunk/frameworks/projects/spark/src/spark/skins/spark/BorderSkin.as

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1286132 13f79535-47bb-0310-9956-ffa450edef68