Transferred from the Adobe Open Source Subversion Repository
Revision: 4978
Author:   gosmith@adobe.com
Date:     2009-02-17T02:25:10.529022Z

Log Message: 
-----------
Various changes to Gumbo text styles.

The marginLeft, marginRight, marginTop, and marginBottom styles have been removed. They used to inherit from a TLF container to a TLF paragraph and got applied there, but this is not longer the case so that TLF can support container-level margins in the future. Since container-level margins are not currently supported, there should not be a corresponding component-level style. In order to make it possible to specify a format on a container to set inter-paragraph spacing, TLF will add spaceBefore and spaceAfter soon, and Flex will expose these as component styles.

The tabStops style is no longer supported on TextBox or components that use TextBox in their skins. It is supported on TextGraphic, TextView, and FxTextArea, but only as a TLF-parsed String (rather than an Array or Vector of TabStop instances as in FTE). Examples of tabstop strings are "10 20 30 40" and "s10 c20 r30 d40|," (where s="start", c="center", r="right", and d="decimal". The use of strings to specify tabstops makes it easy to set them in CSS.

The firstLineOffsetBasis style has been removed. There is no such thing in TLF's TextLayoutFormat any more.

The dominantBaseline, justificationRule, and justificationStyle styles now allow the value "auto", and this is now the default.

The ligatureLevel style no longer allows the value "none".

QE Notes: None
Doc Notes: None
Bugs: None
Reviewer: Ryan

Modified Paths:
--------------
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flex4/src/mx/styles/metadata/AdvancedTextLayoutFormatStyles.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/framework/defaults.css</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flex4/src/mx/graphics/TextBox.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flex4/src/mx/styles/metadata/BasicTextLayoutFormatStyles.as</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1287846 13f79535-47bb-0310-9956-ffa450edef68