Transferred from the Adobe Open Source Subversion Repository
Revision: 16344
Author:   dloverin@adobe.com
Date:     2010-05-27T13:42:03.619655Z

Log Message: 
-----------
Don't remove RSLs that are required for their inheritance dependencies

We can’t remove an RSL that seems unused if it supplies inheritance dependencies for RSLs that will be loaded later.

Modified the order of RSLs in flex-config to put the least used RSLs at the end of the list if possible.  This will reduce the number of downstream RSLs that need to be checked by the compiler. 

QE notes: The textLayout RSL made be pulled into some apps where it was not in previous tests. The Spark RSL has inheritance dependencies on the textLayout RSL.
Doc notes: None
Bugs: SDK-26471
Reviewer: Corey
Tests run: checktests, cyclone server
Is noteworthy for integration: No

Modified Paths:
--------------
<path kind="file" action="M">/flex/sdk/trunk/modules/compiler/src/java/flex2/tools/PreLink.java</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/flex-config.xml</path>
<path kind="file" action="M">/flex/sdk/trunk/modules/compiler/src/java/flex2/tools/oem/Configuration.java</path>
<path kind="file" action="M">/flex/sdk/trunk/modules/compiler/src/java/flex2/compiler/swc/SwcDependencySet.java</path>
<path kind="file" action="M">/flex/sdk/trunk/modules/compiler/src/java/flex2/compiler/util/SwcDependencyUtil.java</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1290064 13f79535-47bb-0310-9956-ffa450edef68