Transferred from the Adobe Open Source Subversion Repository
Revision: 5558
Author:   pfarland@adobe.com
Date:     2009-03-25T21:50:18.008985Z

Log Message: 
-----------
Add hack to Grammar.jj to allow special component tags in spark and halo namespaces to resolve to special node in the parser. This will be re-addressed once we stabilize on the renames and can better enforce the contents of MXML 2006, MXML 2009, FXG, Spark and Halo.

Add a work around for the step building the AIR Updater UI to set -compatibility-version=3.0.0. This involved checking in a local copy to override the sample-frameworks-build.xml checked into the AIR Integration Kit.zip.

Adding missing base halo and spark classes that appear as qualified type selectors in our defaults.css files but have historically not been in our manifests for MXML tag mappings.

QE: This should help pass more tests.
Dev: Not yet
Doc: Not yet

Modified Paths:
--------------
<path kind="file" action="M">/flex/sdk/trunk/frameworks/build.xml</path>
<path kind="file" action="M">/flex/sdk/trunk/modules/compiler/src/java/flex2/compiler/mxml/MXMLNamespaces.java</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/spark-manifest.xml</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/halo-manifest.xml</path>
<path kind="file" action="M">/flex/sdk/trunk/modules/compiler/src/java/flex2/compiler/mxml/dom/MxmlScanner.java</path>
<path kind="file" action="M">/flex/sdk/trunk/modules/compiler/src/java/flex2/compiler/mxml/Grammar.jj</path>
<path kind="file" action="A">/flex/sdk/trunk/in/air/sample-frameworks-build.xml</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1290150 13f79535-47bb-0310-9956-ffa450edef68